row_consumer: de-virtualize io_priority() and resource_tracker()
Fixes #2830

Message-Id: <448a1f739ab8c88a7a5562bce8dce5ae6efdf934.1507302530.git.bdenes@scylladb.com>
